Rename IssueDetection model selection components to GenAI prefix
Renames IssueDetectionModelSelection, IssueDetectionApiKeyConfigurator,
and IssueDetectionAdvancedSettings (and their sub-types/interfaces) to
GenAIModelSelection, GenAIApiKeyConfigurator, and GenAIAdvancedSettings
for reuse across workflows beyond issue detection.
